requirements/base.txt
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.22.0
beautifulsoup4==4.6.0
bleach==2.0.0
cffi==1.10.0
contextlib2==0.5.5
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.10.7
django-ckeditor==5.2.2
django-colorful==1.2
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filebrowser-no-grappelli==3.7.0
django-geoposition==0.3.0
django-import-export==0.5.1
django-indexer==0.3.0
django-jsonview==1.0.0
django-localeurl==2.0.2
django-modeladmin-reorder==0.2
django-modelcluster==3.1
django-modeltranslation==0.12.1
django-nested-admin==3.0.17
django-paging==0.2.5
django-recaptcha==1.3.0
django-taggit==0.22.1
django-templatetag-sugar==1.0
django-treebeard==4.1.0
django-uuslug==1.1.8
docutils==0.13.1
ecdsa==0.13
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
greenlet==0.4.12
html5lib==0.999999999
httplib2==0.10.3
idna==2.5
lxml==3.7.3
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.5.0
pyparsing==2.2.0
python-monkey-business==1.0.0
python-slugify==1.2.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
raven==6.0.0
requests==2.14.2
requests-oauthlib==0.8.0
simplejson==3.10.0
singlemodeladmin==0.7
six==1.10.0
soaplib==1.0.0
social-auth-core==1.3.0
spyne==2.12.14
sqlparse==0.2.3
tablib==0.11.4
Unidecode==0.4.20
webencodings==0.5.1
Willow==0.4
xlrd==1.0.0

